WIP: CUDA wheels

+++ wrapping/dockcross-manylinux-download-cache.sh
@@ -3,16 +3,31 @@
 # This module should be pulled and run from an ITKModule root directory to generate the Linux python wheels of this module,
 # it is used by the circle.yml file contained in ITKModuleTemplate: https://github.com/InsightSoftwareConsortium/ITKModuleTemplate
 
-curl https://data.kitware.com/api/v1/file/592dd8068d777f16d01e1a92/download -o zstd-1.2.0-linux.tar.gz
-gunzip -d zstd-1.2.0-linux.tar.gz
-tar xf zstd-1.2.0-linux.tar
+if [[ ! -f zstd-1.2.0-linux.tar.gz ]]; then
+  curl https://data.kitware.com/api/v1/file/592dd8068d777f16d01e1a92/download -o zstd-1.2.0-linux.tar.gz
+  gunzip -d zstd-1.2.0-linux.tar.gz
+  tar xf zstd-1.2.0-linux.tar
+fi
+if [[ ! -f ./zstd-1.2.0-linux/bin/unzstd ]]; then
+  echo "ERROR: Cannot find required binary './zstd-1.2.0-linux/bin/unzstd'"
+  exit 255
+fi
 
-curl -L https://github.com/InsightSoftwareConsortium/ITKPythonBuilds/releases/download/${ITK_PACKAGE_VERSION:=v5.3rc01}/ITKPythonBuilds-linux.tar.zst -O
+if [[ ! -f iITKPythonBuilds-linux.tar.zst ]]; then
+  curl -L https://github.com/InsightSoftwareConsortium/ITKPythonBuilds/releases/download/${ITK_PACKAGE_VERSION:=v5.3rc01}/ITKPythonBuilds-linux.tar.zst -O
+fi
+if [[ ! -f ./ITKPythonBuilds-linux.tar.zst ]]; then
+  echo "ERROR: cannot find required ${ITK_PACKAGE_VERSION} binary './ITKPythonBuilds-linux.tar.zst'"
+  exit 255
+fi
 ./zstd-1.2.0-linux/bin/unzstd ITKPythonBuilds-linux.tar.zst -o ITKPythonBuilds-linux.tar
+echo "Extracting all files"
 tar xf ITKPythonBuilds-linux.tar
+rm ITKPythonBuilds-linux.tar
 
 mkdir tools
 curl https://data.kitware.com/api/v1/file/5c0aa4b18d777f2179dd0a71/download -o doxygen-1.8.11.linux.bin.tar.gz
 tar -xvzf doxygen-1.8.11.linux.bin.tar.gz -C tools
 
+echo `(ls ITKPythonPackage)`
 cp -a ITKPythonPackage/oneTBB-prefix ./
